Now all Connect modules (Fulfillment, Tier Configuration and Usage) are separated in services with custom service provider for each one. Callable shortcuts has been added to keep backwards compatibility.
