+++ src/AutomationEngine.php
@@ -8,12 +8,23 @@
 
 namespace Connect;
 
+use Connect\Modules\Fulfilment;
+use Connect\Modules\TierConfiguration;
+use Connect\Modules\Usage;
+use GuzzleHttp\ClientInterface;
 use Pimple\Container;
 use Pimple\Psr11\Container as PSRContainer;
 use Psr\Container\ContainerInterface;
+use Psr\Log\LoggerInterface;
 
 /**
  * Class AutomationEngine
+ * @property Config $config
+ * @property LoggerInterface $logger
+ * @property ClientInterface $http
+ * @property Fulfilment $fulfillment
+ * @property TierConfiguration $tierConfiguration
+ * @property Usage $usage
  * @package Connect
  */
 abstract class AutomationEngine
@@ -23,6 +34,17 @@ abstract class AutomationEngine
      * @var ContainerInterface
      */
     private $_container;
+
+    /**
+     * List of services with shortcuts for old implementations
+     * @internal
+     * @var array
+     */
+    private $_shorcuts = [
+        'fulfillment',
+        'tierConfiguration',
+        'usage',
+    ];
 
     /**
      * FulfillmentAutomation constructor.
@@ -61,5 +83,20 @@ abstract class AutomationEngine
         return ($this->_container->has($id))
             ? $this->_container->get($id)
             : null;
+    }
+
+    /**
+     * Dynamically call connect native module functions.
+     * @param string $name
+     * @param array $arguments
+     * @return mixed
+     */
+    public function __call($name, $arguments)
+    {
+        foreach ($this->_shorcuts as $id) {
+            if (is_callable([$this->_container->get($id), $name])) {
+                return call_user_func_array([$this->_container->get($id), $name], $arguments);
+            }
+        }
     }
 }
