+++ tests/Unit/ConfigTest.php
@@ -64,7 +64,7 @@ class ConfigTest extends \Test\TestCase
         $cfg = new Config(__DIR__ . '/cfg.valid.runtimeservices.json');
         $this->assertInstanceOf('\Connect\Config', $cfg);
         $this->assertInternalType('array', $cfg->runtimeServices);
-        $this->assertCount(3, $cfg->runtimeServices);
+        $this->assertCount(6, $cfg->runtimeServices);
     }
 
     /**
