+++ tests/Unit/Runtime/Providers/Http4TierConfigServiceProvider.php
@@ -29,7 +29,7 @@ class Http4TierConfigServiceProvider extends ServiceProvider
         $response->shouldReceive('getBody')
             ->andReturn($body);
 
-        $client = \Mockery::mock('Client');
+        $client = \Mockery::mock('GuzzleHttp\ClientInterface');
         $client->shouldReceive('request')
             ->withAnyArgs()
             ->andReturn($response);
