+++ tests/Unit/Runtime/Providers/Http4UsageAutomationServiceProvider.php
@@ -28,7 +28,7 @@ class Http4UsageAutomationServiceProvider extends ServiceProvider
         $response->shouldReceive('getBody')
             ->andReturn($body);
 
-        $client = \Mockery::mock('Client');
+        $client = \Mockery::mock('GuzzleHttp\ClientInterface');
         $client->shouldReceive('request')
             ->withAnyArgs()
             ->andReturn($response);
