+++ tests/Unit/Runtime/Providers/Http4UsageAutomationTestsServiceProvider.php
@@ -29,7 +29,7 @@ class Http4UsageAutomationTestsServiceProvider extends ServiceProvider
         $response->shouldReceive('getBody')
             ->andReturn($body);
 
-        $client = \Mockery::mock('Client');
+        $client = \Mockery::mock('GuzzleHttp\ClientInterface');
         $client->shouldReceive('request')
             ->withAnyArgs()
             ->andReturn($response);
