+++ tests/Unit/Runtime/Providers/HttpFileAutomationServiceProvider.php
@@ -25,7 +25,7 @@ class HttpFileAutomationServiceProvider extends ServiceProvider
         $response->shouldReceive('getBody')
             ->andReturn($body);
 
-        $client = \Mockery::mock('Client');
+        $client = \Mockery::mock('GuzzleHttp\ClientInterface');
         $client->shouldReceive('request')
             ->withAnyArgs()
             ->andReturn($response);
