+++ tests/Unit/Runtime/Providers/LoggerServiceProvider.php
@@ -2,8 +2,6 @@
 
 namespace Test\Unit\Runtime\Providers;
 
-use Connect\Config;
-use Connect\Logger;
 use Connect\Runtime\ServiceProvider;
 use Pimple\Container;
 
@@ -15,7 +13,7 @@ class LoggerServiceProvider extends ServiceProvider
 {
     public function register(Container $container)
     {
-        $logger = \Mockery::mock('LoggerInterface');
+        $logger = \Mockery::mock('Psr\Log\LoggerInterface');
 
         $logger->shouldReceive('emergency')->withAnyArgs();
         $logger->shouldReceive('alert')->withAnyArgs();
